requirements.txt
telethon>=1.43
python-dotenv>=1.0
google-genai>=1.73
litellm>=1.50
